+++ src/main/java/com/notenoughmail/tfcgenviewer/util/ClientHandoff.java
@@ -21,7 +21,10 @@ public class ClientHandoff {
                     Permissions.canSeeCoordinates(perms),
                     Permissions.canSeeSeed(perms),
                     Units.blockToGrid(player.getBlockX()),
-                    Units.blockToGrid(player.getBlockZ())
+                    Units.blockToGrid(player.getBlockZ()),
+                    info.visualizableFeatures(),
+                    info.biomeInfo(),
+                    info.biomeTags()
             ));
         }
     }