+++ src/main/java/com/notenoughmail/tfcgenviewer/util/preview/Image.java
@@ -3,6 +3,7 @@ package com.notenoughmail.tfcgenviewer.util.preview;
 import com.mojang.blaze3d.platform.NativeImage;
 import com.notenoughmail.tfcgenviewer.TFCGenViewer;
 import com.notenoughmail.tfcgenviewer.mixin.NativeImageAccessor;
+import com.notenoughmail.tfcgenviewer.util.MutableImage;
 import net.minecraft.client.renderer.texture.DynamicTexture;
 import net.minecraft.util.FastColor;
 import net.minecraftforge.fml.loading.FMLPaths;
@@ -16,7 +17,7 @@ import java.nio.file.Path;
  * <p>
  * Also provides a number of helpful utils for safely writing to the image
  */
-public class Image {
+public class Image implements MutableImage {
 
     /**
      * Gets the internal {@link NativeImage} of the image. Preserves the nullness of the passed image
@@ -34,7 +35,8 @@ public class Image {
         image = new NativeImage(size, size, false);
     }
 
-    private boolean inRange(int z) {
+    @Override
+    public boolean inRange(int z) {
         return z >= 0 && z <= maxPixel;
     }
 
@@ -54,6 +56,7 @@ public class Image {
         }
     }
 
+    @Override
     public void setPixel(int x, int y, int abgrColor) {
         if (!image.isOutsideBounds(x, y)) {
             final int alpha = FastColor.ABGR32.alpha(abgrColor);
@@ -65,6 +68,7 @@ public class Image {
         }
     }
 
+    @Override
     public void hLine(int x0, int x1, int y, int width, int abgrColor) {
         final int alpha = FastColor.ABGR32.alpha(abgrColor);
         if (alpha != 0) {
@@ -100,6 +104,7 @@ public class Image {
         }
     }
 
+    @Override
     public void vLine(int y0, int y1, int x, int width, int abgrColor) {
         final int alpha = FastColor.ABGR32.alpha(abgrColor);
         if (alpha != 0) {
@@ -135,16 +140,31 @@ public class Image {
         }
     }
 
+    @Override
+    public int getABGRColor(int x, int y) {
+        if (!image.isOutsideBounds(x, y)) {
+            synchronized (image) {
+                if (isAllocated()) {
+                    return image.getPixelRGBA(x, y);
+                }
+            }
+        }
+        return 0;
+    }
+
+    @Override
     public void close() {
         synchronized (image) {
             image.close();
         }
     }
 
+    @Override
     public boolean isAllocated() {
         return ((NativeImageAccessor) (Object) image).tfcgenviewer$GetPixels() != 0L;
     }
 
+    @Override
     public void export(String name) {
         try {
             image.writeToFile(new File(FMLPaths.getOrCreateGameRelativePath(Path.of("screenshots", "tfcgenviewer")).toFile(), name));
@@ -153,6 +173,8 @@ public class Image {
         }
     }
 
+    // This has a problem where while the image may be allocated when *requesting* the upload,
+    // the image can be unallocated afterward, before the image is uploaded on the render thread
     public void upload(DynamicTexture tex) {
         synchronized (image) {
             if (isAllocated()) {